split avro.util
